react decompose
