+++ README.md
@@ -3,6 +3,6 @@
 - Follow the [React task guideline](https://github.com/mate-academy/react_task-guideline#react-tasks-guideline)
 
 ## Task
-Split the `App.jsx` into smaller components basing on CSS blocks. CSS should be split too.
+Split the `App.js` into smaller components basing on CSS blocks. CSS should be split too.
 - Create a sub folder in `components` per component and put `js` and `css` files there.
 - There should be 3 - 5 new components.
