+++ src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import './App.css';
-import Welcome from '../Welcome/Welcome';
-import Header from '../Header/Header';
-import Article from '../Article/Article';
+import Welcome from './components/Welcome/Welcome';
+import Header from './components/Header/Header';
+import Article from './components/Article/Article';
 
 function App() {
   return (
