--- src/components/App/App.css
@@ -1,34 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-html {
-  font-size: 10pt;
-  line-height: 1.4;
-  font-weight: 400;
-  font-family: "Source Sans Pro", "Open Sans", Roboto, Arial, sans-serif;
-}
-
-body {
-  margin: 0;
-  background: #eee;
-}
-
-@media (min-width: 500px) {
-  .header {
-    text-align: left;
-    display: flex;
-    justify-content: space-between;
-  }
-
-  .article {
-    margin: 3.5em auto 0;
-    padding: 2em;
-  }
-}
-
-@media (min-width: 800px) {
-  .article {
-    margin: 3.5em auto;
-  }
-}
