+++ src/components/Article/Article.css
@@ -15,3 +15,16 @@
 .article__paragraph:not(:first-child) {
   margin-top: 0.7em;
 }
+
+@media (min-width: 500px) {
+  .article {
+    margin: 3.5em auto 0;
+    padding: 2em;
+  }
+}
+
+@media (min-width: 800px) {
+  .article {
+    margin: 3.5em auto;
+  }
+}
