+++ src/components/Header/Header.css
@@ -17,3 +17,11 @@
   padding: 0;
   font-size: inherit;
 }
+
+@media (min-width: 500px) {
+  .header {
+    text-align: left;
+    display: flex;
+    justify-content: space-between;
+  }
+}
